+++ Editor/ObjectMapping/ObjectMappingBuilder.cs
@@ -54,35 +54,43 @@ namespace Anatawa12.AvatarOptimizer
             _beforeGameObjectInfos[rootObject.GetInstanceID()].InitializeRecursive();
             
 #if AAO_VRM0
-            if (rootObject.TryGetComponent<VRMFirstPerson>(out var firstPerson))
+            if (rootObject.TryGetComponent<VRMFirstPerson>(out var firstPerson)
+                && firstPerson is { Renderers: { } vrmRenderers} )
             {
-                foreach (var renderer in firstPerson.Renderers)
+                foreach (var renderer in vrmRenderers)
                 {
-                    GetComponentInfo(renderer.Renderer).VrmFirstPersonFlag = renderer.FirstPersonFlag switch
+                    if (renderer.Renderer is { } rendererComponent)
                     {
-                        FirstPersonFlag.Auto => VrmFirstPersonFlag.Auto,
-                        FirstPersonFlag.Both => VrmFirstPersonFlag.Both,
-                        FirstPersonFlag.ThirdPersonOnly => VrmFirstPersonFlag.ThirdPersonOnly,
-                        FirstPersonFlag.FirstPersonOnly => VrmFirstPersonFlag.FirstPersonOnly,
-                        _ => throw new ArgumentOutOfRangeException()
-                    };
+                        GetComponentInfo(rendererComponent).VrmFirstPersonFlag = renderer.FirstPersonFlag switch
+                        {
+                            FirstPersonFlag.Auto => VrmFirstPersonFlag.Auto,
+                            FirstPersonFlag.Both => VrmFirstPersonFlag.Both,
+                            FirstPersonFlag.ThirdPersonOnly => VrmFirstPersonFlag.ThirdPersonOnly,
+                            FirstPersonFlag.FirstPersonOnly => VrmFirstPersonFlag.FirstPersonOnly,
+                            _ => throw new ArgumentOutOfRangeException()
+                        };
+                    }
                 }
             }
 #endif
 
 #if AAO_VRM1
-            if (rootObject.TryGetComponent<Vrm10Instance>(out var vrm10Instance))
+            if (rootObject.TryGetComponent<Vrm10Instance>(out var vrm10Instance)
+                && vrm10Instance is { Vrm.FirstPerson.Renderers: { } vrm10Renderers })
             {
-                foreach (var renderer in vrm10Instance.Vrm.FirstPerson.Renderers)
+                foreach (var renderer in vrm10Renderers)
                 {
-                    GetComponentInfo(renderer.GetRenderer(rootObject.transform)).VrmFirstPersonFlag = renderer.FirstPersonFlag switch
+                    if (renderer.GetRenderer(rootObject.transform) is { } rendererComponent)
                     {
-                        FirstPersonType.auto => VrmFirstPersonFlag.Auto,
-                        FirstPersonType.both => VrmFirstPersonFlag.Both,
-                        FirstPersonType.thirdPersonOnly => VrmFirstPersonFlag.ThirdPersonOnly,
-                        FirstPersonType.firstPersonOnly => VrmFirstPersonFlag.FirstPersonOnly,
-                        _ => throw new ArgumentOutOfRangeException()
-                    };
+                        GetComponentInfo(rendererComponent).VrmFirstPersonFlag = renderer.FirstPersonFlag switch
+                        {
+                            FirstPersonType.auto => VrmFirstPersonFlag.Auto,
+                            FirstPersonType.both => VrmFirstPersonFlag.Both,
+                            FirstPersonType.thirdPersonOnly => VrmFirstPersonFlag.ThirdPersonOnly,
+                            FirstPersonType.firstPersonOnly => VrmFirstPersonFlag.FirstPersonOnly,
+                            _ => throw new ArgumentOutOfRangeException()
+                        };
+                    }
                 }
             }
 #endif