+++ CHANGELOG.md
@@ -37,7 +37,7 @@ The format is based on [Keep a Changelog].
   - Merge ToonLit Material will be removed in next major version.
   - This component will support both Skinned Mesh Renderer and Mesh Renderer.
 - Replace EndBone With Endpoint Position component which replaces the end bone in the vrc physbone with the Endpoint Position [`#1423`](https://github.com/anatawa12/AvatarOptimizer/pull/1423)
-- Automatic Replace EndBone With Endpoint Position [`#1423`](https://github.com/anatawa12/AvatarOptimizer/pull/1423)
+- Automatic Replace EndBone With Endpoint Position `#1423` `#1604`
   - Trace and Optimize now automatically replaces PhysBone's EndBone with Endpoint Position if possible.
 - Experimental support for NDMF Platform Support `#1579` `#1577` `#1576`
   - This is an experimental feature that does not follow semantic versioning.