+++ CHANGELOG.md
@@ -111,6 +111,10 @@ The format is based on [Keep a Changelog].
 
 ### Security
 
+## [1.8.16] - 2025-12-03
+### Changed
+- Streaming mipmap settings are copied when processing textures [`#1558`](https://github.com/anatawa12/AvatarOptimizer/pull/1558)
+
 ## [1.8.15] - 2025-11-24
 ### Added
 - Support for VRCSDK 3.10.x [`#1562`](https://github.com/anatawa12/AvatarOptimizer/pull/1562) [`#1570`](https://github.com/anatawa12/AvatarOptimizer/pull/1570)
@@ -1347,7 +1351,8 @@ The format is based on [Keep a Changelog].
 - Merge Bone
 - Clear Endpoint Position
 
-[Unreleased]: https://github.com/anatawa12/AvatarOptimizer/compare/v1.8.15...HEAD
+[Unreleased]: https://github.com/anatawa12/AvatarOptimizer/compare/v1.8.16...HEAD
+[1.8.16]: https://github.com/anatawa12/AvatarOptimizer/compare/v1.8.15...v1.8.16
 [1.8.15]: https://github.com/anatawa12/AvatarOptimizer/compare/v1.8.14...v1.8.15
 [1.8.14]: https://github.com/anatawa12/AvatarOptimizer/compare/v1.8.13...v1.8.14
 [1.8.13]: https://github.com/anatawa12/AvatarOptimizer/compare/v1.8.12...v1.8.13